+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project are documented here. The format is
 based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 this project adheres to [Semantic Versioning](https://semver.org/).
+## [unreleased]
+
+### Features
+
+- *(auth)* Make OIDC scopes configurable, drop groups from the default
+
 ## [0.8.3] - 2026-09-02
 
 ### Bug Fixes